Added events

+++ src/services/FlushCacheService.php
@@ -7,6 +7,7 @@ namespace putyourlightson\blitz\services;
 
 use Craft;
 use craft\base\Component;
+use putyourlightson\blitz\events\RefreshCacheEvent;
 use putyourlightson\blitz\models\SiteUriModel;
 use putyourlightson\blitz\records\CacheRecord;
 use putyourlightson\blitz\records\ElementQueryRecord;
@@ -18,44 +19,17 @@ class FlushCacheService extends Component
     // =========================================================================
 
     /**
-     * @event Event
+     * @event RefreshCacheEvent
+     */
+    const EVENT_BEFORE_FLUSH_CACHE = 'beforeFlushCache';
+
+    /**
+     * @event RefreshCacheEvent
      */
     const EVENT_AFTER_FLUSH_CACHE = 'afterFlushCache';
 
     // Public Methods
     // =========================================================================
-
-    /**
-     * Flushes the entire cache.
-     */
-    public function flushAll()
-    {
-        CacheRecord::deleteAll();
-
-        $this->runGarbageCollection();
-
-        // Fire an 'afterFlushCache' event
-        if ($this->hasEventHandlers(self::EVENT_AFTER_FLUSH_CACHE)) {
-            $this->trigger(self::EVENT_AFTER_FLUSH_CACHE);
-        }
-    }
-
-    /**
-     * Flushes the cache for a given site.
-     *
-     * @param int $siteId
-     */
-    public function flushSite(int $siteId)
-    {
-        CacheRecord::deleteAll(['siteId' => $siteId]);
-
-        $this->runGarbageCollection();
-
-        // Fire an 'afterFlushCache' event
-        if ($this->hasEventHandlers(self::EVENT_AFTER_FLUSH_CACHE)) {
-            $this->trigger(self::EVENT_AFTER_FLUSH_CACHE);
-        }
-    }
 
     /**
      * Flushes cache records given an array of site URIs.
@@ -64,7 +38,13 @@ class FlushCacheService extends Component
      */
     public function flushUris(array $siteUris)
     {
-        foreach ($siteUris as $siteUri) {
+        $event = $this->onBeforeFlush(['siteUris' => $siteUris]);
+
+        if (!$event->isValid) {
+            return;
+        }
+
+        foreach ($event->siteUris as $siteUri) {
             CacheRecord::deleteAll([
                 'siteId' => $siteUri->siteId,
                 'uri' => $siteUri->uri,
@@ -75,7 +55,51 @@ class FlushCacheService extends Component
 
         // Fire an 'afterFlushCache' event
         if ($this->hasEventHandlers(self::EVENT_AFTER_FLUSH_CACHE)) {
-            $this->trigger(self::EVENT_AFTER_FLUSH_CACHE);
+            $this->trigger(self::EVENT_AFTER_FLUSH_CACHE, $event);
+        }
+    }
+
+    /**
+     * Flushes the cache for a given site.
+     *
+     * @param int $siteId
+     */
+    public function flushSite(int $siteId)
+    {
+        $event = $this->onBeforeFlush(['siteId' => $siteId]);
+
+        if (!$event->isValid) {
+            return;
+        }
+
+        CacheRecord::deleteAll(['siteId' => $siteId]);
+
+        $this->runGarbageCollection();
+
+        // Fire an 'afterFlushCache' event
+        if ($this->hasEventHandlers(self::EVENT_AFTER_FLUSH_CACHE)) {
+            $this->trigger(self::EVENT_AFTER_FLUSH_CACHE, $event);
+        }
+    }
+
+    /**
+     * Flushes the entire cache.
+     */
+    public function flushAll()
+    {
+        $event = $this->onBeforeFlush();
+
+        if (!$event->isValid) {
+            return;
+        }
+
+        CacheRecord::deleteAll();
+
+        $this->runGarbageCollection();
+
+        // Fire an 'afterFlushCache' event
+        if ($this->hasEventHandlers(self::EVENT_AFTER_FLUSH_CACHE)) {
+            $this->trigger(self::EVENT_AFTER_FLUSH_CACHE, $event);
         }
     }
 
@@ -101,6 +125,27 @@ class FlushCacheService extends Component
         if (ElementQueryRecord::find()->count() == 0) {
             $this->_resetAutoIncrement(ElementQueryRecord::tableName());
         }
+    }
+
+    // Protected Methods
+    // =========================================================================
+
+    /**
+     * Triggers the on before event.
+     *
+     * @param array $config
+     *
+     * @return RefreshCacheEvent
+     */
+    protected function onBeforeFlush(array $config)
+    {
+        $event = new RefreshCacheEvent($config);
+
+        if ($this->hasEventHandlers(self::EVENT_BEFORE_FLUSH_CACHE)) {
+            $this->trigger(self::EVENT_BEFORE_FLUSH_CACHE, $event);
+        }
+
+        return $event;
     }
 
     // Private Methods
